add teachable module
