+++ apprentice_teacher.rb
@@ -1,18 +1,16 @@
 require_relative 'school_class'
+require_relative 'teachable'
 
 class ApprenticeTeacher < School
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
+  include Teachable
 
   def initialize(options={})
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
     @target_raise = 800
     @phase = 3
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -28,10 +26,6 @@ class ApprenticeTeacher < School
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
 
   def receive_raise(raise)
     @salary += raise
