+++ school_class.rb
@@ -8,7 +8,7 @@
 # STUDENT, TEACHER and APPRENTICE TEACHER all have the following shared METHODS:
 
   # offer_high_five
-  # set_phase
+  # set_phase => Note this brings up an error since the response is different between teachers and stu
 
 # Don't forget about CONSTANTS yo... => TEACHER_MAX_HEATH = 100 => self.class::TEACHER_MAX_HEALTH
 
@@ -17,27 +17,12 @@ class School
   attr_accessor :name
 
   def initialize
-        @age = options.fetch(:age, 0)
+    @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
   end
 
-   def offer_high_five
+  def offer_high_five
     "High five!"
   end
-
-  def set_phase(num)
-    response = ""
-    if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
-    else
-      response = "Oooh, phase #{num}. I hope I'm ready!"
-    end
-    @phase = num
-    response
-  end
-
-
-
 
 end
