+++ student.rb
@@ -10,7 +10,17 @@ class Student < School
     @name = options.fetch(:name, "")
   end
 
-
+  def set_phase(num)
+    response = ""
+    if num == @phase
+      response += "I'm doing phase #{@phase} again because "
+      response += "I put my learning first. I'm gonna rock it!"
+    else
+      response = "Oooh, phase #{num}. I hope I'm ready!"
+    end
+    @phase = num
+    response
+  end
 
   def learn_stuff
     response = ""
