+++ teacher.rb
@@ -1,9 +1,12 @@
 require_relative 'school_class'
+require_relative 'teachable'
+
 
 
 class Teacher < School
   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
   attr_accessor :name
+  include Teachable
 
   def initialize(options={})
     @phase = 3
@@ -12,9 +15,7 @@ class Teacher < School
     @target_raise = 1000
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # note the SET_PHASE and TEACH_STUFF method is slightly different so they cant be refactored?
 
   def set_phase(num)
     @phase = num
@@ -27,15 +28,6 @@ class Teacher < School
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
